Changed teleport to use relative coordinates

+++ src/class.rs
@@ -20,7 +20,7 @@ impl Class {
             skills: sksk,
         }
     }
-    pub fn use_handle(x: i32, y: i32, now: Duration) -> impl Fn(&mut GameState) -> bool {
+    pub fn use_handle(x: f64, y: f64, now: Duration) -> impl Fn(&mut GameState) -> bool {
         move |gs: &mut GameState| -> bool {
             let n = match &gs.scene {
                 Scenes::GamePlay(c) => {match &c {
@@ -33,7 +33,7 @@ impl Class {
             if n < gs.class.as_ref().unwrap().skills.len() {
                 let mut gd = gs.gamedata.lock().unwrap();
                 println!("ok so rn this is just using the coords of where you click with the mouse for this jonathicc u need to actually convert into rrelative cooords pls :)))){} {}",x,y);
-                return gs.class.as_mut().unwrap().skills[n].useskill(&mut gd, Position {x: x as f64, y: y as f64}, now)
+                return gs.class.as_mut().unwrap().skills[n].useskill(&mut gd, Position {x: x, y: y}, now)
             }
             false
         }
